xiyan_mcp_server/pyproject.toml
[project]
name = "xiyan_mcp_server"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server that using XiyanSQL with MySQL databases. This server allows AI assistants to list tables, read data, and execute natual language queries."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "mysql-connector-python>=9.1.0",
    "llama_index",
    "sqlalchemy"
]
[[project.authors]]
name = "Zhiling Luo"
email = "godot.lzl@alibaba-inc.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mysql_mcp_server = "xiyan_mcp_server:main"